se valido entero en cantidad
.

+++ database/migrations/2022_10_09_190004_create_finanzas_table.php
@@ -29,7 +29,7 @@ class CreateFinanzasTable extends Migration
             $table->dateTime('fecha_entrada')->nullable();
             $table->dateTime('fecha_facturacion')->nullable();
             $table->string('descripcion');
-            $table->decimal('cantidad', 12, 2);
+            $table->integer('cantidad');
             $table->integer('unidad_id');
             $table->decimal('costo_unitario', 12, 2);
             $table->decimal('monto_a_pagar', 12, 2);
